Use named return err param for callback, events handlers moved to separated file
