+++ packages/client-proxy/internal/edge-pass-through/proxy.go
@@ -98,7 +98,7 @@ func (s *NodePassThroughServer) director(ctx context.Context) (*grpc.ClientConn,
 //
 // Core implementation is just following methods that are handling forwarding, proper closing and propagating of errors from both sides of the stream.
 // The handler is called for every request that is not handled by any other gRPC service.
-func (s *NodePassThroughServer) handler(srv interface{}, serverStream grpc.ServerStream) error {
+func (s *NodePassThroughServer) handler(srv interface{}, serverStream grpc.ServerStream) (err error) {
 	fullMethodName, ok := grpc.MethodFromServerStream(serverStream)
 	if !ok {
 		return status.Errorf(codes.Internal, "low lever server stream not exists in context")
@@ -125,19 +125,18 @@ func (s *NodePassThroughServer) handler(srv interface{}, serverStream grpc.Serve
 	clientCtx, clientCancel := context.WithCancel(s.ctx)
 	defer clientCancel()
 
+	clientStream, err := grpc.NewClientStream(clientCtx, clientStreamDescForProxying, clientConnection, fullMethodName)
+	if err != nil {
+		return err
+	}
+
 	callback, err := s.eventsHandler(md)
 	if err != nil {
 		return status.Errorf(codes.Internal, "failed to handle events: %v", err)
 	}
 
-	var returnErr error
 	if callback != nil {
-		defer callback(returnErr)
-	}
-
-	clientStream, err := grpc.NewClientStream(clientCtx, clientStreamDescForProxying, clientConnection, fullMethodName)
-	if err != nil {
-		return err
+		defer callback(err)
 	}
 
 	// Explicitly *do not close* s2cErrChan and c2sErrChan, otherwise the select below will not terminate.
@@ -159,7 +158,6 @@ func (s *NodePassThroughServer) handler(srv interface{}, serverStream grpc.Serve
 				// to cancel the clientStream to the backend, let all of its goroutines be freed up by the CancelFunc and
 				// exit with an error to the stack
 				clientCancel()
-				returnErr = s2cErr
 				return status.Errorf(codes.Internal, "failed proxying s2c: %v", s2cErr)
 			}
 		case c2sErr := <-c2sErrChan:
@@ -169,7 +167,6 @@ func (s *NodePassThroughServer) handler(srv interface{}, serverStream grpc.Serve
 			serverStream.SetTrailer(clientStream.Trailer())
 			// c2sErr will contain RPC error from client code. If not io.EOF return the RPC error as server stream error.
 			if c2sErr != io.EOF {
-				returnErr = c2sErr
 				return c2sErr
 			}
 			return nil
@@ -229,71 +226,4 @@ func (s *NodePassThroughServer) forwardServerToClient(src grpc.ServerStream, dst
 	}()
 
 	return ret
-}
-
-func (s *NodePassThroughServer) eventsHandler(md metadata.MD) (func(error), error) {
-	eventTypeHeaders := md.Get(edge.EventTypeHeader)
-	if len(eventTypeHeaders) == 0 {
-		return nil, nil
-	}
-
-	if len(eventTypeHeaders) > 1 {
-		return nil, status.Errorf(codes.InvalidArgument, "multiple event types are not supported: %v", eventTypeHeaders)
-	}
-
-	eventType := eventTypeHeaders[0]
-	switch eventType {
-	case edge.CatalogCreateEventType:
-		return s.catalogCreateEventHandler(md)
-	case edge.CatalogDeleteEventType:
-		return s.catalogDeleteEventHandler(md)
-	default:
-		return nil, status.Errorf(codes.InvalidArgument, "event type %s is not supported", eventType)
-	}
-}
-
-func (s *NodePassThroughServer) catalogCreateEventHandler(md metadata.MD) (func(error), error) {
-	c, err := edge.ParseSandboxCatalogCreateEvent(md)
-	if err != nil {
-		return nil, err
-	}
-
-	err = s.catalog.StoreSandbox(
-		c.SandboxID,
-		&sandboxes.SandboxInfo{
-			OrchestratorID:          c.OrchestratorID,
-			ExecutionID:             c.ExecutionID,
-			SandboxStartedAt:        c.SandboxStartTime,
-			SandboxMaxLengthInHours: c.SandboxMaxLengthInHours,
-		},
-		time.Duration(c.SandboxMaxLengthInHours)*time.Hour,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("failed to store sandbox in catalog: %w", err)
-	}
-
-	return func(err error) {
-		if err == nil {
-			return
-		}
-
-		deleteErr := s.catalog.DeleteSandbox(c.SandboxID, c.ExecutionID)
-		if deleteErr != nil {
-			zap.L().Error("Failed to delete sandbox in catalog after failing request", zap.Error(deleteErr))
-		}
-	}, nil
-}
-
-func (s *NodePassThroughServer) catalogDeleteEventHandler(md metadata.MD) (func(error), error) {
-	d, err := edge.ParseSandboxCatalogDeleteEvent(md)
-	if err != nil {
-		return nil, err
-	}
-
-	err = s.catalog.DeleteSandbox(d.SandboxID, d.ExecutionID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to delete sandbox from catalog: %w", err)
-	}
-
-	return nil, nil
 }
